requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cleanup==8.0.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.6.0
